+++ examples/LilBomber/lil_bomber_gun1.lua
@@ -1,11 +1,11 @@
 meta = {
     name = "Lil bomber",
     version = "WIP",
-    author = "Estebanfer",
+    author = "SereneRuby12",
     description = ""
 }
 
-local c_ent_lib = import("estebanfer/custom-entities-library")
+local c_ent_lib = import("SereneRuby12/custom-entities-library")
 --local c_ent_lib = require "custom_entities"
 
 local bomb_spawn_sound = get_sound(VANILLA_SOUND.ENEMIES_OLMEC_BOMB_SPAWN)